refactor(extensions): organize default extensions into ui and vcs tiers
The compiled-in tier now lives under src/extensions/default/, split by what it
contributes: default/vcs/{git,jujutsu,sapling} for the VCS backends — each with
its full name and its own directory — and default/ui/sidebar for the built-in
sidebar. The directory layout now mirrors the tier split the loaders already
enforce: default/vcs is what VCS adapter resolution imports and must stay
renderer-free, while default/ui loads from the app's sidebar resolution.
Extension ids (git, jj, sl) and registration behavior are unchanged.
