+++ AGENTS.md
@@ -27,7 +27,7 @@ CLI input
 - Pager mode has two paths: full diff UI for patch-like stdin, plain-text fallback for non-diff pager content.
 - View defaults are layered through built-ins, user config, repo `.hunk/config.toml`, command sections, pager sections, and CLI flags.
 - `hunk daemon serve` runs one loopback daemon that brokers agent commands to many live Hunk sessions. Normal Hunk sessions should auto-start and register with that daemon when session brokering is enabled. Keep it local-only and session-brokered rather than opening per-TUI ports.
-- Extensions come in two tiers. User TypeScript extensions (`src/extensions/`) load at interactive-app startup, before `loadAppBootstrap`; **every VCS backend Hunk ships — Git, jj, and Sapling — is a bundled extension** in `src/extensions/bundled/`, statically imported and loaded synchronously from VCS adapter resolution so they are available during config resolution too. There are zero core-registered VCS adapters: Git registers through the public `registerVcsAdapter` like everything else, which is what keeps the published API honest, because Git is the backend that exercises every integration point. Both tiers run through the same per-extension API object and registry collection (`src/extensions/runExtension.ts`). Bundled extensions are implicitly trusted and stay loaded under `--no-extensions`, which governs user extensions only. Registrations (themes, file languages, VCS adapters, changeset transforms, sidebar views, lifecycle events) are collected into one registry and applied through `src/extensions/apply.ts` on both startup and reload. Extension files import `react` and `hunkdiff/extension` as host-served runtime modules (`src/extensions/hostRuntimeModules.ts`), which is what lets `registerSidebarView` components render inside the app's own React tree with working hooks. The built-in file-navigation sidebar is itself a bundled extension (`src/extensions/bundled/sidebar/`): it registers through `registerSidebarView` and consumes exactly the public sidebar props, which is what keeps that contract sufficient for third-party sidebars; a user-registered sidebar view overrides the bundled default. It is deliberately not part of `loadBundledExtensions` — that list loads from VCS adapter resolution, which must stay renderer-free — and instead loads through `getBundledSidebarView` where the app resolves its active sidebar. Repo-local `.hunk/extensions/` is trust-gated: never execute repo code without the trust prompt. Git keeps the detection-priority baseline (`HUNK_CORE_VCS_DETECTION_PRIORITY = 0`, jj 200 / sl 100 above it); `src/core/vcs/index.ts` is the single assembly point that orders bundled + user adapters by `detectionPriority`. Auto-detection is uniform across both tiers: nearest checkout wins, `detectionPriority` breaks equal-distance ties, and an explicit `vcs` id a loaded backend owns always beats detection. `src/extensions/vcsPatchResult.ts` is the one conversion boundary where a published `ExtensionVcsPatchResult` becomes Hunk's internal diff model — exact file sources, extra/skipped files, and user-facing errors all cross there, so anything a backend needs that cannot be expressed publicly is a real gap in the contract. The public authoring surface is the `hunkdiff/extension` export — keep it a façade over internal types; authoring guide in `docs/extensions.md`. That contract is declared in `src/extension-api/types.ts`, which must stay import-free: declaration emission ships every module the entry reaches, so an import there publishes Hunk internals (and `scripts/check-pack.ts` fails the pack when it does). Shapes shared with internal code are declared there and re-exported inward.
+- Extensions come in two tiers. User TypeScript extensions (`src/extensions/`) load at interactive-app startup, before `loadAppBootstrap`; **every VCS backend Hunk ships — Git, jj, and Sapling — is a bundled extension** in `src/extensions/default/vcs/`, statically imported and loaded synchronously from VCS adapter resolution so they are available during config resolution too. There are zero core-registered VCS adapters: Git registers through the public `registerVcsAdapter` like everything else, which is what keeps the published API honest, because Git is the backend that exercises every integration point. Both tiers run through the same per-extension API object and registry collection (`src/extensions/runExtension.ts`). Bundled extensions are implicitly trusted and stay loaded under `--no-extensions`, which governs user extensions only. Registrations (themes, file languages, VCS adapters, changeset transforms, sidebar views, lifecycle events) are collected into one registry and applied through `src/extensions/apply.ts` on both startup and reload. Extension files import `react` and `hunkdiff/extension` as host-served runtime modules (`src/extensions/hostRuntimeModules.ts`), which is what lets `registerSidebarView` components render inside the app's own React tree with working hooks. The built-in file-navigation sidebar is itself a bundled extension (`src/extensions/default/ui/sidebar/`): it registers through `registerSidebarView` and consumes exactly the public sidebar props, which is what keeps that contract sufficient for third-party sidebars; a user-registered sidebar view overrides the bundled default. It is deliberately not part of `loadBundledExtensions` — that list loads from VCS adapter resolution, which must stay renderer-free — and instead loads through `getBundledSidebarView` where the app resolves its active sidebar. Repo-local `.hunk/extensions/` is trust-gated: never execute repo code without the trust prompt. Git keeps the detection-priority baseline (`HUNK_CORE_VCS_DETECTION_PRIORITY = 0`, jj 200 / sl 100 above it); `src/core/vcs/index.ts` is the single assembly point that orders bundled + user adapters by `detectionPriority`. Auto-detection is uniform across both tiers: nearest checkout wins, `detectionPriority` breaks equal-distance ties, and an explicit `vcs` id a loaded backend owns always beats detection. `src/extensions/vcsPatchResult.ts` is the one conversion boundary where a published `ExtensionVcsPatchResult` becomes Hunk's internal diff model — exact file sources, extra/skipped files, and user-facing errors all cross there, so anything a backend needs that cannot be expressed publicly is a real gap in the contract. The public authoring surface is the `hunkdiff/extension` export — keep it a façade over internal types; authoring guide in `docs/extensions.md`. That contract is declared in `src/extension-api/types.ts`, which must stay import-free: declaration emission ships every module the entry reaches, so an import there publishes Hunk internals (and `scripts/check-pack.ts` fails the pack when it does). Shapes shared with internal code are declared there and re-exported inward.
 - Agent rationale is optional sidecar JSON matched onto files/hunks.
 - The order of `files` in the sidecar is intentional. Hunk uses that order for the sidebar and main review stream.
 - Prefer one source of truth for each user-visible behavior. When rendering, navigation, scrolling, or note placement share the same model, derive them from the same planning layer rather than maintaining parallel implementations.
