+++ src/extensions/default/ui/sidebar/index.test.tsx
