+++ src/extensions/default/vcs/git/index.test.ts
@@ -2,13 +2,13 @@ import { afterEach, describe, expect, test } from "bun:test";
 import { mkdirSync, mkdtempSync, realpathSync, rmSync, writeFileSync } from "node:fs";
 import { platform, tmpdir } from "node:os";
 import { isAbsolute, join, resolve } from "node:path";
-import { GitVcsAdapter, statSignature } from "./git";
+import { GitVcsAdapter, statSignature } from ".";
 import type {
   ExtensionVcsDiffInput,
   ExtensionVcsOperations,
   ExtensionVcsShowInput,
   ExtensionVcsStashShowInput,
-} from "../../extension-api/types";
+} from "../../../../extension-api/types";
 
 // The adapter is written against the published contract, so the tests read it
 // through that contract too — including the capabilities Git is the only
